refactor(electromag): creates TPZScalarField base class
