+++ Material/Electromagnetics/TPZPeriodicWgma.h
@@ -1,9 +1,8 @@
 #ifndef _TPZPLANARWGMODALANALYSIS_H_
 #define _TPZPLANARWGMODALANALYSIS_H_
 
-#include "TPZMatBase.h"
+#include "TPZScalarField.h"
 #include "TPZMatGeneralisedEigenVal.h"
-#include "TPZMatSingleSpace.h"
 
 /**
  * @ingroup material
@@ -18,57 +17,21 @@
  Yasuhide Tsuji and Masanori Koshiba, "Finite Element Method Using Port Truncation by Perfectly Matched Layer Boundary Conditions for Optical Waveguide Discontinuity Problems," J. Lightwave Technol. 20, 463- (2002) 
 */
 class TPZPeriodicWgma
-    : public TPZMatBase<CSTATE, TPZMatSingleSpaceT<CSTATE>,
-                        TPZMatGeneralisedEigenVal> {
-  using TBase =
-      TPZMatBase<CSTATE, TPZMatSingleSpaceT<CSTATE>, TPZMatGeneralisedEigenVal>;
+  : public TPZScalarField,
+    public TPZMatGeneralisedEigenVal {
 
 public:
-  //! Sets for which mode type the problem will be solved.
-  enum class ModeType { TE, TM };
-  /**
-     @brief Constructor taking a few material parameters
-     @param[in] id Material identifier.
-     @param[in] dim Integrable dimension 
-     @param[in] er Relative permittivity.
-     @param[in] ur Relative permeability.
-     @param[in] scale Scale for geometric domain.
-     @note the `scale` param might help with floating point arithmetics on
-     really small domains.
-  */
-  TPZPeriodicWgma(int id, int dim, const CSTATE er, const CSTATE ur,
-                           const STATE lambda, const ModeType mode,
-                           const REAL &scale = 1.);
 
+  //! Parent class's constructor
+  using TPZScalarField::TPZScalarField;
+  
   TPZPeriodicWgma *NewMaterial() const override;
 
   std::string Name() const override { return "TPZPeriodicWgma"; }
 
-  /** @brief Returns the integrable dimension of the material */
-  int Dimension() const override { return fDim; }
-
-  [[nodiscard]] int NStateVariables() const override { return 1; }
-
-  /**
-     @name ParamMethods
-     @{
-  */
-  //! Sets the wavelength being analysed
-  void SetWavelength(STATE lambda);
-  //! Gets the current wavelength
-  [[nodiscard]] inline STATE GetWavelength() const { return fLambda; }
-  //! Sets the permeability of the material
-  void SetPermeability(const CSTATE ur);
-  //! Gets the permeability of the material
-  virtual void GetPermeability(
-    [[maybe_unused]] const TPZVec<REAL> &x,TPZVec<CSTATE> &ur) const;
-  //! Sets the permittivity of the material
-  void SetPermittivity(const CSTATE er);
-  //! Gets the permittivity of the material
-  virtual void GetPermittivity(
-    [[maybe_unused]] const TPZVec<REAL> &x,TPZVec<CSTATE> &er) const;
   //! Sets the propagation constant to be used in the K matrix
   void SetBeta(const CSTATE beta){fBeta = beta;}
+  
   /**@}*/
   /**
      @name ContributeMethods
@@ -81,16 +44,25 @@ public:
                     TPZFMatrix<CSTATE> &ek, TPZFMatrix<CSTATE> &ef,
                     TPZBndCondT<CSTATE> &bc) override;
   /**@}*/
-  /**
-     @name SolutionMethods
-     @{*/
-  //! Variable index of a given solution
-  int VariableIndex(const std::string &name) const override;
-  //! Number of variables associated with a given solution
-  int NSolutionVariables(int var) const override;
-  //! Computes the solution at an integration point
-  void Solution(const TPZMaterialDataT<CSTATE> &data, int var,
-                TPZVec<CSTATE> &solout) override;
+
+  [[nodiscard]] int ClassId() const override;
+    
+  void Read(TPZStream& buf, void* context) override;
+
+  void Write(TPZStream& buf, int withclassid) const override;
+
+
+  TPZBndCondT<CSTATE>* CreateBC(TPZMaterial *reference,
+                               int id, int type,
+                               const TPZFMatrix<CSTATE> &val1,
+                               const TPZVec<CSTATE> &val2) override
+    {
+      return new  TPZBndCondBase<CSTATE,
+                                 TPZMatSingleSpaceBC<CSTATE>,
+                                 TPZMatGeneralisedEigenValBC>
+        (reference,id, type,val1,val2);
+    }
+  
 protected:
   //! Actual Contribute method (both TE/TM) for matrix A
   void ContributeInternalA(const CSTATE cGradX, const CSTATE cGradY,
@@ -107,23 +79,11 @@ protected:
                           const CSTATE cScal,
                           const TPZMaterialDataT<CSTATE> &data, REAL weight,
                           TPZFMatrix<CSTATE> &ek, TPZFMatrix<CSTATE> &ef);
-  //! Mode type being solved
-  ModeType fMode{ModeType::TE};
-  //! Relative magnetic permeability
-  CSTATE fUr{1};
-  //! Relative electric permittivity
-  CSTATE fEr{1};
-  //! Wavelength being analysed
-  STATE fLambda{1.55e-9};
-  //! Scale factor for the domain (helps with floating point arithmetic on small
-  //! domains)
-  const REAL fScaleFactor{1.};
-  //! Integrable dimension
-  const int fDim{2};
+  
   //! Propagation constant
   CSTATE fBeta;
   //! Default constructot
-  TPZPeriodicWgma();
+  TPZPeriodicWgma() = default;
 };
 
 #endif /* _TPZPLANARWGMODALANALYSIS_H_ */
